Exception handling

+++ src/main/java/com/solucaotecnologia/TechCommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.solucaotecnologia.TechCommerce.services;
 import com.solucaotecnologia.TechCommerce.dto.ProductDTO;
 import com.solucaotecnologia.TechCommerce.entities.Product;
 import com.solucaotecnologia.TechCommerce.repositories.ProductRepository;
+import com.solucaotecnologia.TechCommerce.services.exceptions.DatabaseException;
+import com.solucaotecnologia.TechCommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -22,13 +27,9 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
         Optional<Product> result = productRepository.findById(id);
-        if (result.isPresent()) {
-            Product product = result.get();
-            ProductDTO productDTO = new ProductDTO(product);
-            return productDTO;
-        } else {
-            throw new RuntimeException("Produto não encontrado");
-        }
+        Product product = result.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
+        ProductDTO productDTO = new ProductDTO(product);
+        return productDTO;
     }
 
     @Transactional(readOnly = true)
@@ -48,19 +49,31 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product product = productRepository.getReferenceById(id);
-        product.setName(dto.getName());
-        product.setDescription(dto.getDescription());
-        product.setPrice(dto.getPrice());
-        product = productRepository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            product.setName(dto.getName());
+            product.setDescription(dto.getDescription());
+            product.setPrice(dto.getPrice());
+            product = productRepository.save(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
+
 
 
 }
